+++ src/app/providers/app-config.service.ts
@@ -1,14 +1,16 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '../../environments/environment';
+// services
+import { Globals } from '../utils/globals';
 
 @Injectable()
 export class AppConfigService {
   private appConfig;
 
-  constructor(private http: HttpClient) {
+  constructor(private http: HttpClient, public g: Globals) {
     this.appConfig = environment;
-    console.log('environment: ', environment);
+    this.g.wdLog(['environment: ', environment]);
   }
 
   loadAppConfig() {